perf/rplayer_library:

- Shrink librplayer.so to about 3MB for armv7-a & armv8-a devices using FFmpeg build script from Bilibili/ijkplayer.
- Fix texture id overflow on armv7-a devices.

+++ src/ffi/ffi_export.h
@@ -42,7 +42,7 @@ int RPlayer_getState(void*);
 
 char* RPlayer_getMessage(void*);
 
-long RPlayer_getTextureId(void*);
+long long RPlayer_getTextureId(void*);
 
 //////////////////////////////////////////
 // RPlayerConfig
