+++ src/ffi/ffi_flutter.cpp
@@ -58,12 +58,12 @@ DART_EXPORT char* RPlayer_getMessage(void* p) {
   return pPlayer->msg;
 }
 
-DART_EXPORT long RPlayer_getTextureId(void* p) {
+DART_EXPORT long long RPlayer_getTextureId(void* p) {
   RPlayer* pPlayer = static_cast<RPlayer*>(p);
   if (pPlayer->pTextureAndroid == nullptr) {
     return -1;
   }
-  return static_cast<long>(pPlayer->pTextureAndroid->id);
+  return pPlayer->pTextureAndroid->id;
 }
 
 //////////////////////////////////////////
